Add initial version of vue project
To develop, open JsonRpc.Docs/web in the editor and run 'npm install' and then 'npm run serve' to serve the page.
The development version reads a file from the public folder, while the production reads the generated file (they have the same path).
To use it in the ASP.NET project, run 'npm run build' then compile the ASP.NET project, the web/dist is the folder which contains the static resources.