Add dialog to add a server to the list of servers
The dialog isn't functional yet, it is just the form with some basic validation.
The idea is to store the locally added servers in the local storage and also remember the user last choice.
The user should also be allowed to remove locally added servers (the remove button should show up when a locally added server is selected)